pyproject.toml
[project]
name = "survey-medley-code"
version = "0.1.0"
description = "Survey Medley analysis code"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.7",
    "nibabel>=5.3.2",
    "nilearn>=0.12.1",
    "numba>=0.62.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pyyaml>=6.0.3",
    "randomise-prep",
    "rpy2>=3.6.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "typing>=3.10.0.0",
    "vif-for-contrasts",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
randomise-prep = { git = "https://github.com/jmumford/randomise-prep", rev = "6c73254" }
vif-for-contrasts = { git = "https://github.com/jmumford/vif_contrasts.git" }


[tool.ruff]
extend-include = ["*.ipynb"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
extend-select = ["E501", "W", "N", "R", "I"]
ignore = ["F401"]  # allow unused imports

[tool.ruff.lint.pydocstyle]
convention = "numpy"
